createMapSelector: basic implementation & tests
